Creation of the backend
creation of the web api backend for the to do's app. It has three projects, the Data that has entity framework and interact with sqlite, the core and the Api itself with the controller.
Apart from that initiation of the frontend with reach, still some needed changes, middle of progress